portfolio-fullstack/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
python-dotenv>=1.0.1

